module-1-langchain/requirements.txt
langchain
langchain-openai
langchain-google
langchain-google-genai
pydantic
python-dotenv
langgraph
wikipedia
langchain-community
argparse